requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
supabase==2.3.4
python-dotenv==1.0.0
urllib3==2.1.0
